Id type fixed, List items by logged in user.

+++ src/main/java/mk/com/theagrodiarybackend/model/Expense.java
@@ -14,7 +14,7 @@ public class Expense {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long expenseId;
+    private Integer expenseId;
     @Column(nullable = false)
     private Float expenseSum;
     private ZonedDateTime date;
