+++ src/main/java/mk/com/theagrodiarybackend/model/dto/PlantationDto.java
@@ -11,6 +11,6 @@ public class PlantationDto {
     private Float amountKg;
     private String type;
     private Integer year;
-    private Long personId;
-    private Long seedId;
+    private Integer personId;
+    private Integer seedId;
 }
