+++ src/main/java/mk/com/theagrodiarybackend/model/exception/ExpenseNotFoundException.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class ExpenseNotFoundException extends RuntimeException {
 
-    public ExpenseNotFoundException(Long expenseId) {
+    public ExpenseNotFoundException(Integer expenseId) {
         super(String.format("Expense with id: %d not found!", expenseId));
     }
 }
