+++ src/main/java/mk/com/theagrodiarybackend/repository/SeedRepository.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SeedRepository extends JpaRepository<Seed, Long> {
+public interface SeedRepository extends JpaRepository<Seed, Integer> {
 
     @Query("select s from Seed s where s.seedName = :seedName")
     Optional<Seed> findBySeedName(String seedName);
     @Query("select s from Seed s where s.seedId = :seedId")
-    Optional<Seed> findBySeedId(Long seedId);
+    Optional<Seed> findBySeedId(Integer seedId);
 }
